requirements.txt
kafka-python==2.2.16
